Create 4_14_Third_Commit.md
I added the code for save and quit and deleted the code for show_all. I think the function for show_all is unnecessary. I will continue to do my work, and I submitted this earlier afternoon on Sunday. If I make any changes this evening, I will upload it.